% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HealthIneq.R
\docType{package}
\name{HealthIneq}
\alias{HealthIneq}
\alias{Health}
\alias{HealthIneq-package}
\title{Health Inequities}
\format{
A data frame with 2600 elements
\describe{
\item{cz}{Commuting Zone ID}
\item{czname}{Commuting Zone Name}
\item{pop2000}{Commuting Zone Population in 2000}
\item{fips}{State FIPS}
\item{statename}{State Name}
\item{stateabbrv}{State Abbreviation}
\item{puninsured2010}{Percent Uninsured}
\item{reimb_penroll_adj10}{Medicare $ Per Enrollee}
\item{rel_tot}{Percent Religious}
\item{hhinc00}{Mean Household Income}
}
}
\source{
\url{https://healthinequality.org/data/}
}
\usage{
HealthIneq
}
\description{
This package contains data on US counties
pertaining to life expectancy.

A data set containing information on
health inequities across the United
States by county.
}
\keyword{datasets}
